feat: implement multi-language support with new pages and navigation components
