+++ app/routes/_public+/($lang)._index/components/CoverPage.tsx
