+++ app/routes/_public+/($lang)._index/components/Navigation.tsx
