+++ app/routes/_public+/($lang)._index/components/pages/index.ts
